test(parquet/metadata): keep schema error coverage with file metadata
